Starting to add Download page functionality
